+++ src/Commands/Backup/CreateCommand.php
@@ -2,18 +2,12 @@
 
 namespace Pantheon\Terminus\Commands\Backup;
 
-use Pantheon\Terminus\Commands\TerminusCommand;
-use Pantheon\Terminus\Site\SiteAwareInterface;
-use Pantheon\Terminus\Site\SiteAwareTrait;
-
 /**
  * Class CreateCommand
  * @package Pantheon\Terminus\Commands\Backup
  */
-class CreateCommand extends TerminusCommand implements SiteAwareInterface
+class CreateCommand extends BackupCommand
 {
-    use SiteAwareTrait;
-
     /**
      * Creates a backup of a specific site and environment.
      *
@@ -22,7 +16,7 @@ class CreateCommand extends TerminusCommand implements SiteAwareInterface
      * @command backup:create
      *
      * @param string $site_env Site & environment in the format `site-name.env`
-     * @option string $element [code|files|database|db] Element to be backed up. If not defined, all elements are selected.
+     * @option string $element [all|code|files|database|db] Element to be backed up
      * @option integer $keep-for Retention period, in days, to retain backup
      *
      * @usage <site>.<env> Creates a backup of <site>'s <env> environment.
@@ -30,12 +24,10 @@ class CreateCommand extends TerminusCommand implements SiteAwareInterface
      * @usage <site>.<env> --keep-for=<days> Creates a backup of <site>'s <env> environment and retains it for <days> days.
      * @usage <site>.<env> --keep-for=<days> Creates a backup of <site>'s <env> environment's <element> and retains it for <days> days.
      */
-    public function create($site_env, $options = ['element' => null, 'keep-for' => 365,])
+    public function create($site_env, $options = ['element' => 'all', 'keep-for' => 365,])
     {
         list(, $env) = $this->getSiteEnv($site_env);
-        if (isset($options['element']) && ($options['element'] == 'db')) {
-            $options['element'] = 'database';
-        }
+        $options['element'] = isset($options['element']) ? $this->getElement($options['element']) : null;
         $env->getBackups()->create($options)->wait();
         $this->log()->notice('Created a backup of the {env} environment.', ['env' => $env->id,]);
     }