+++ src/Commands/Backup/GetCommand.php
@@ -2,21 +2,17 @@
 
 namespace Pantheon\Terminus\Commands\Backup;
 
-use Pantheon\Terminus\Commands\TerminusCommand;
 use Pantheon\Terminus\Request\RequestAwareInterface;
 use Pantheon\Terminus\Request\RequestAwareTrait;
-use Pantheon\Terminus\Site\SiteAwareInterface;
-use Pantheon\Terminus\Site\SiteAwareTrait;
 use Pantheon\Terminus\Exceptions\TerminusNotFoundException;
 
 /**
  * Class GetCommand
  * @package Pantheon\Terminus\Commands\Backup
  */
-class GetCommand extends TerminusCommand implements SiteAwareInterface, RequestAwareInterface
+class GetCommand extends BackupCommand implements RequestAwareInterface
 {
     use RequestAwareTrait;
-    use SiteAwareTrait;
 
     /**
      * Displays the download URL for a specific backup or latest backup.
@@ -27,7 +23,7 @@ class GetCommand extends TerminusCommand implements SiteAwareInterface, RequestA
      *
      * @param string $site_env Site & environment in the format `site-name.env`
      * @option string $file [filename.tgz] Name of backup file
-     * @option string $element [code|files|database|db] Backup element to retrieve. If not defined, all elements are selected.
+     * @option string $element [all|code|files|database|db] Backup element to retrieve
      * @option string $to Local path to save to
      * @throws TerminusNotFoundException
      *
@@ -37,14 +33,14 @@ class GetCommand extends TerminusCommand implements SiteAwareInterface, RequestA
      * @usage <site>.<env> --to=<path> Saves the most recent backup of any type in <site>'s <env> environment to <path>.
      * @usage <site>.<env> --to=<path> Saves the most recent <element> backup in <site>'s <env> environment to <path>.
      */
-    public function getBackup($site_env, array $options = ['file' => null, 'element' => null, 'to' => null,])
+    public function getBackup($site_env, array $options = ['file' => null, 'element' => 'all', 'to' => null,])
     {
         list($site, $env) = $this->getSiteEnv($site_env);
 
         if (isset($options['file']) && !is_null($file_name = $options['file'])) {
             $backup = $env->getBackups()->getBackupByFileName($file_name);
         } else {
-            $element = ($options['element'] == 'db') ? 'database' : $options['element'];
+            $element = isset($options['element']) ? $this->getElement($options['element']) : null;
             $backups = $env->getBackups()->getFinishedBackups($element);
             if (empty($backups)) {
                 throw new TerminusNotFoundException(